Address Copilot review feedback
- Recreate points_ledger index after table rebuild in migration
- Handle 0-point routines with bonus in push notification body
- Simplify insertBonusIfPositive with early returns
- Extract shared useAdminSettings hook, reuse in useAdminTimezone
